+++ CHANGELOG.md
@@ -5,6 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0] - 05-08-2025
+
+### Added
+
+- `HebrewDate` now holds `include_festive_days` and `include_fasts` as attributes that can be changed after
+   instantiation.
+- `HebrewDate` is now iterable, yielding the `day_numeric`, `month_numeric` and `year_numeric` attributes.
+
+### Changed
+
+- ⚠️ **IMPORTANT:** Changed `HebrewDate` initialization behavior: 
+    If both day and month are passed, then year must also be passed.
+    If neither day, month nor year are passed, then the current date will be used,
+    which makes the default constructor equivalent to `HebrewDate.today()`.
+    Finally, if only day or month are not passed, then they are set to the first day or the first month respectively.
+- methods: `HebrewDate.get_month`, `HebrewDate.get_day` and `HebrewDate.get_weekday` got suffixed with `_tuple`
+- `HebrewDate.is_holiday` and `HebrewDate.holiday` are now dynamic properties instead of attributes
+
+### Fixed
+
+- Validation logic for `HebrewDate`
+
 ## [2.0.4] - 27-06-2025
 
 ### Changed